Selectable query in alert

+++ src/components/layers/layer.js
@@ -8,6 +8,8 @@ import Accordion from 'react-bootstrap/Accordion';
 import Card from 'react-bootstrap/Card';
 import { Form, Button } from 'react-bootstrap';
 import { CypherEditor } from "graph-app-kit/components/Editor"
+import { confirmAlert } from 'react-confirm-alert'; // Import
+import 'react-confirm-alert/src/react-confirm-alert.css'; // Import css
 
 // css needed for CypherEditor
 import "codemirror/lib/codemirror.css";
@@ -311,12 +313,26 @@ class Layer extends Component {
     deleteLayer(event) {
 	/*Remove the layer
 	 */
-	this.props.deleteLayer(this.state.ukey);
 	event.preventDefault();
+	if (
+	    window.confirm(
+		`Delete layer ${this.state.name}? This action can not be undone.`
+	    ) === false
+	) {
+	    return;
+	}
+	this.props.deleteLayer(this.state.ukey);
     };
 
     showQuery(event) {
-	alert(this.getQuery());
+	confirmAlert({
+	    message: this.getQuery(),
+	    buttons: [
+		{
+		    label: 'OK',
+		}
+	    ]
+	});
 	event.preventDefault();
     };
